Unify gateway appender into istio_details.go
It's not worth to have a separate appender to decorate ingress gateway nodes. So, unifying the code.

Also, use the same pattern as in ServiceEntry nodes to show configured hostnames.
