+++ graph/telemetry/istio/appender/appender.go
@@ -25,8 +25,6 @@ func ParseAppenders(o graph.TelemetryOptions) []graph.Appender {
 				requestedAppenders[AggregateNodeAppenderName] = true
 			case DeadNodeAppenderName:
 				requestedAppenders[DeadNodeAppenderName] = true
-			case GatewayHostnamesAppenderName:
-				requestedAppenders[GatewayHostnamesAppenderName] = true
 			case HealthConfigAppenderName:
 				requestedAppenders[HealthConfigAppenderName] = true
 			case IdleNodeAppenderName:
@@ -163,10 +161,6 @@ func ParseAppenders(o graph.TelemetryOptions) []graph.Appender {
 		a := SidecarsCheckAppender{
 			AccessibleNamespaces: o.AccessibleNamespaces,
 		}
-		appenders = append(appenders, a)
-	}
-	if _, ok := requestedAppenders[GatewayHostnamesAppenderName]; ok || o.Appenders.All {
-		a := GatewayHostnamesAppender{}
 		appenders = append(appenders, a)
 	}
